backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
psycopg2-binary
pydantic
passlib
python-jose
requests
httpx
google-genai
google
google-auth
tenacity
anyio
h11
httpcore
idna
charset-normalizer
certifi
urllib3
websockets
typing-extensions